+++ ts/main.ts
@@ -1,26 +1,22 @@
 import { Listener } from './api/stream/listener'
 import { Stream } from './api/stream/stream'
 import { API } from './bot/api'
+import { playOXGame, resetOXGame } from './bot/oxgame'
 import { randomContent } from './botcontents'
 import { Auth, Configure } from './conf'
 import { evalCalc } from './evalcalc'
 import { rePattern, sholdWipeTL } from './filter/repattern'
 import { filterWords } from './filter/secret'
-import { OXGame } from './oxgame/gamestate'
 import { findCoordinate } from './oxgame/input'
 import { tootParser } from './tootparser'
 import { IArgumentToot } from './types/apitype'
 import { INotification, IParsedToot, IStatus } from './types/deftype'
-import { Coordinate, Mark } from './types/oxgametype'
 
 const bot = {
   // username@example.com
   ID: '',
   username: ''
 }
-
-// OXGame states.
-const oxGameStates: {[key: string]: OXGame} = {}
 
 const getParsedToot = (toot: IStatus): IParsedToot => ({
   account: tootParser.screenName(toot.account),
@@ -149,68 +145,6 @@ const wipeTL = (toot: IParsedToot): void => {
     visibility: toot.visibility
   }
   setTimeout(() => API.write.toot(sendData), 6000)
-}
-
-const playOXGame = (toot: IParsedToot, oxCoordinate: Coordinate | null, mark: Mark, ismention?: boolean): void => {
-  // Setup
-  if (!oxGameStates[toot.account]) {
-    const newGame = new OXGame(mark)
-    oxGameStates[toot.account] = newGame
-  }
-
-  const state = oxGameStates[toot.account]
-  let result = null
-  if (oxCoordinate) result = state.move(oxCoordinate)
-  else result = state.initMove()
-  const nowState = state.stateView()
-  if (result === '◯' || result === '✕' || result === 'draw') {
-    delete oxGameStates[toot.account]
-  } else if (result === 'invalid') {
-    const invalid = 'そこゎ置けないょ(∩´﹏`∩)💦'
-    const sendDataOnInvalid: IArgumentToot = {
-      status: `@${toot.account} ${invalid}`,
-      visibility: toot.visibility
-    }
-    if (ismention) sendDataOnInvalid.in_reply_to_id = toot.id
-    setTimeout(() => API.write.toot(sendDataOnInvalid), 3000)
-    return
-  }
-
-  const playerMark = state.player
-  const botMark = state.bot
-
-  let prefixMsg = ''
-  switch (result) {
-    // Player wins.
-    case playerMark:
-      prefixMsg = randomContent.oxGameYouWin()
-      break
-    case botMark:
-      prefixMsg = randomContent.oxGameYouLose()
-      break
-    case null:
-      prefixMsg = randomContent.oxGameThinking()
-      break
-    case 'draw':
-      prefixMsg = '引き分けだね〜(๑>◡<๑)'
-      break
-  }
-
-  const msg = `${prefixMsg}\n${nowState}`
-  const sendData: IArgumentToot = {
-    spoiler_text: `あなた: ${playerMark} ぅゅ: ${botMark}`,
-    status: `@${toot.account} ${msg}`,
-    visibility: toot.visibility
-  }
-  if (ismention) sendData.in_reply_to_id = toot.id
-  setTimeout(() => API.write.toot(sendData), 3000)
-}
-
-const resetOXGame = (toot: IParsedToot, oxCoordinate: Coordinate | null, mark: Mark, ismention?: boolean): void => {
-  const newGame = new OXGame(mark)
-  oxGameStates[toot.account] = newGame
-
-  playOXGame(toot, oxCoordinate, mark, ismention)
 }
 
 const close = (toot: IParsedToot): void => {
